+++ cardano-base/src/Cardano/Base/Bytes.hs
@@ -2,8 +2,11 @@
 {-# LANGUAGE TypeApplications #-}
 
 module Cardano.Base.Bytes (
+  byteArrayFromByteString,
   byteArrayToByteString,
   byteStringToByteArray,
+  byteArrayFromShortByteString,
+  byteArrayToShortByteString,
   slice,
   splitsAt,
 )
@@ -22,9 +25,14 @@ byteArrayToByteString :: ByteArray -> ByteString
 byteArrayToByteString = SBS.fromShort . byteArrayToShortByteString
 {-# INLINE byteArrayToByteString #-}
 
+byteArrayFromByteString :: ByteString -> ByteArray
+byteArrayFromByteString = byteArrayFromShortByteString . SBS.toShort
+{-# INLINE byteArrayFromByteString #-}
+
 byteStringToByteArray :: ByteString -> ByteArray
 byteStringToByteArray = byteArrayFromShortByteString . SBS.toShort
 {-# INLINE byteStringToByteArray #-}
+{-# DEPRECATED byteStringToByteArray "In favor of more consistently named `byteArrayFromByteString`" #-}
 
 slice :: Word -> Word -> ByteString -> ByteString
 slice offset size =