+++ cardano-base/CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog for `cardano-base`
 
-## 0.1.5.1
+## 0.1.6.0
 
-*
+* Re-export `byteArrayFromShortByteString`, `byteArrayToShortByteString`.
+* Deprecate `byteStringToByteArray` in favor of newly added `byteArrayFromByteString`
 
 ## 0.1.5.0