+++ cardano-crypto-class/CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog for `cardano-crypto-class`
 
-## 2.5.0.1
+## 2.5.1.0
 
-*
+* Add `psbToByteArray`
+* Add `psbFromByteStringM`
 
 ## 2.5.0.0